suggested changes and bug fix
