+++ examples/dell5k_edge_system_with_filter_iotcc.py
@@ -39,11 +39,24 @@ from liota.entities.edge_systems.dell5k_edge_system import Dell5KEdgeSystem
 from liota.dcc_comms.websocket_dcc_comms import WebSocketDccComms
 from liota.dccs.dcc import RegistrationFailure
 from liota.lib.utilities.filters.range_filter import RangeFilter, Type
-from liota.lib.utilities.filters.windowed_filters.windowed_range_filter import WindowedRangeFilter
+from liota.lib.utilities.filters.windowing_scheme.windowing_scheme import WindowingScheme
 
 # getting values from conf file
 config = {}
 execfile('sampleProp.conf', config)
+
+# Filters to filter data at Sampling Functions
+# Simple filters
+cpu_pro_filter = RangeFilter(Type.CLOSED_REJECT, 5, 10)  # If no of CPU processes <=5 or >=10
+cpu_util_filter = RangeFilter(Type.AT_LEAST, None, 85)  # CPU util >= 85
+disk_usage_filter = RangeFilter(Type.AT_LEAST, None, 80)  # Disk usage >= 80%
+net_usage_filter = RangeFilter(Type.AT_LEAST, None, 1000000)  # Network usage >= 1Mb
+mem_free_filter = RangeFilter(Type.AT_MOST, 15, None)  # Memory Free <= 15%
+
+# Filters with windowing scheme
+net_usage_filter_with_window = WindowingScheme(net_usage_filter, 30)
+mem_free_filter_with_window = WindowingScheme(mem_free_filter, 60)
+
 
 # some standard metrics for Linux systems
 # agent classes for different IoT system
@@ -52,25 +65,10 @@ execfile('sampleProp.conf', config)
 # we are showing here how to create a representation for a Device in IoTCC but
 # using the notion of RAM (because we have no connected devices yet)
 # agent classes for different kinds of layer 4/5 connections from agent to DCC
-# -------User defined functions for getting the next value for a metric --------
+# -------User defined functions with filters for getting the next value for a metric --------
 # usage of these shown below in main
 # semantics are that on each call the function returns the next available value
 # from the device or system associated to the metric.
-
-# Filters to filter data at Sampling Functions
-# Simple filters
-cpu_pro_filter = RangeFilter(Type.CLOSED_REJECT, 5, 10)  # If no of CPU processes <=5 or >=10
-cpu_util_filter = RangeFilter(Type.AT_LEAST, None, 85)  # CPU util >= 85
-disk_usage_filter = RangeFilter(Type.AT_LEAST, None, 80)  # Disk usage >= 80%
-
-# Filters with windowing scheme
-net_usage_filter = WindowedRangeFilter(Type.AT_LEAST, None, 1000000, 30)  # Network usage >= 1Mb
-mem_free_filter = WindowedRangeFilter(Type.AT_MOST, 15, None, 60)  # Memory Free <= 15%
-
-
-# ---------------------------------------------------------------------------
-# User defined methods with RangeFilters
-
 
 def read_cpu_procs():
     cnt = 0
@@ -91,16 +89,16 @@ def read_disk_usage_stats():
 
 
 def read_network_bytes_received():
-    return net_usage_filter.filter(round(psutil.net_io_counters(pernic=False).bytes_recv, 2))
+    return net_usage_filter_with_window.filter(round(psutil.net_io_counters(pernic=False).bytes_recv, 2))
 
 
 def read_mem_free():
-    return mem_free_filter.filter(round((100 - psutil.virtual_memory().percent), 2))
+    return mem_free_filter_with_window.filter(round((100 - psutil.virtual_memory().percent), 2))
 
 
 # ---------------------------------------------------------------------------
 # In this example, we demonstrate how System health and some simluated data
-# can be directed to two data center components (IoTCC and graphite) using Liota.
+# can be directed to two data center components IoTCC using Liota.
 # The program illustrates the ease of use Liota brings to IoT application developers.
 
 if __name__ == '__main__':
