+++ liota/lib/utilities/filters/windowing_scheme/windowing_scheme.py
@@ -30,38 +30,83 @@
 #  THE POSSIBILITY OF SUCH DAMAGE.                                            #
 # ----------------------------------------------------------------------------#
 
+from abc import abstractmethod
+from numbers import Number
 import logging
 
-from liota.lib.utilities.filters.range_filter import RangeFilter
-from liota.lib.utilities.filters.windowed_filters.windowed_filter import WindowedFilter
+from liota.lib.utilities.utility import getUTCmillis
+from liota.lib.utilities.filters.filter import Filter
 
 log = logging.getLogger(__name__)
 
 
-class WindowedRangeFilter(RangeFilter, WindowedFilter):
+class WindowingScheme(Filter):
     """
-    RangeFilter with windowing scheme.
+    Applies windowing scheme to a Filter.
 
     It keeps track of a configurable time window.  Even if all values has been filtered out at the
     end of every time window, collected value is returned so that DCC is aware of it.
     """
 
-    def __init__(self, filter_type=None, lower_bound=None, upper_bound=None, window_size_sec=10):
+    def __init__(self, filter_obj, window_size_sec):
         """
-        :param filter_type: Any one type from Filter Enum of RangeFilter.
-        :param lower_bound: Lower bound
-        :param upper_bound: Upper bound
-        :param window_size_sec: Configurable time window for heartbeat
+        :param filter_obj: Filter object.
+        :param window_size_sec: Window size in seconds.
         """
-        RangeFilter.__init__(self, filter_type, lower_bound, upper_bound)
-        WindowedFilter.__init__(self, window_size_sec)
+
+        if not isinstance(filter_obj, Filter):
+            raise TypeError("Filter Object is expected")
+        if not isinstance(window_size_sec, Number) or window_size_sec < 0:
+            log.error("window_size_sec must be a non negative number")
+            raise ValueError("window_size_sec must be a non negative number")
+        self.filter_obj = filter_obj
+        self.window_size_sec = window_size_sec
+        #  To track whether at-least one sample has been passed after filtering within a window
+        self.sample_passed = False
+        self.next_window_time = getUTCmillis() + (self.window_size_sec * 1000)
 
     def filter(self, v):
         """
-        Child classes must implement appropriate filtering logic.
+        Applies filter and windowing scheme.
 
         :param v: Collected value by sampling function.
         :return: Filtered value or None
         """
-        log.info("Applying WindowedRangeFilter")
-        return self._window(v, super(WindowedRangeFilter, self).filter(v))
+        log.info("Applying windowing scheme")
+        return self._window(v, self.filter_obj.filter(v))
+
+    def _window(self, collected_value, filtered_value):
+        """
+        Windowing scheme.
+
+        :param collected_value: Collected value by sampling function.
+        :param filtered_value: Filtered value by sampling function.
+        :return: Filtered value (or) collected value at the end of every time window.
+        """
+        # Next window time has elapsed.
+        if getUTCmillis() >= self.next_window_time:
+            #  At-least one sample has not passed so far during this window.
+            if not self.sample_passed and filtered_value is None:
+                self._set_next_window_time()
+                log.info("Sending collected value for this window.")
+                return collected_value
+
+            # At-least one sample has (or will be) passed by now.
+            else:
+                self._set_next_window_time()
+                return filtered_value  # Could be filtered-value or None
+
+        # Next window time has not elapsed.
+        else:
+            if filtered_value is not None:
+                self.sample_passed = True  # At-least one sample has passed during this window
+            return filtered_value
+
+    def _set_next_window_time(self):
+        """
+        Sets next time-window.
+        :return: None
+        """
+        self.next_window_time += (self.window_size_sec * 1000)
+        log.info("Resetting window")
+        self.sample_passed = False  # Resetting
